SoC: update SPI drivers for hbirdv2

+++ SoC/hbirdv2/Common/Include/hbirdv2_spi.h
@@ -18,14 +18,14 @@
 #define SPI_CSN2        2
 #define SPI_CSN3        3
 
-int32_t spi_setup_cmd_addr(SPI_TypeDef *spi, int cmd, int cmdlen, int addr, int addrlen);
-int32_t spi_setup_dummy(SPI_TypeDef *spi, int dummy_rd, int dummy_wr);
-int32_t spi_setup_clk(SPI_TypeDef *spi, int clkdiv);
-int32_t spi_set_datalen(SPI_TypeDef *spi, int datalen);
-int32_t spi_start_transaction(SPI_TypeDef *spi, int trans_type, int csnum);
+int32_t spi_setup_cmd_addr(SPI_TypeDef *spi, uint32_t cmd, uint32_t cmdlen, uint32_t addr, uint32_t addrlen);
+int32_t spi_setup_dummy(SPI_TypeDef *spi, uint16_t dummy_rd, uint16_t dummy_wr);
+int32_t spi_setup_clk(SPI_TypeDef *spi, uint8_t clkdiv);
+int32_t spi_set_datalen(SPI_TypeDef *spi, uint32_t datalen);
+int32_t spi_start_transaction(SPI_TypeDef *spi, uint32_t trans_type, uint32_t csnum);
 int32_t spi_get_status(SPI_TypeDef *spi);
-int32_t spi_write_fifo(SPI_TypeDef *spi, int *data, int datalen);
-int32_t spi_read_fifo(SPI_TypeDef *spi, int *data, int datalen);
+int32_t spi_write_fifo(SPI_TypeDef *spi, uint32_t *data, uint32_t datalen);
+int32_t spi_read_fifo(SPI_TypeDef *spi, uint32_t *data, uint32_t datalen);
 int32_t spi_enable_interrupt(SPI_TypeDef *spi);
 int32_t spi_disable_interrupt(SPI_TypeDef *spi);
 int32_t spi_get_int_status(SPI_TypeDef *spi);
