+++ SoC/hbirdv2/Common/Source/Drivers/hbirdv2_spi.c
@@ -2,7 +2,7 @@
 #include "hbirdv2_spi.h"
 
 
-int32_t spi_setup_cmd_addr(SPI_TypeDef *spi, int cmd, int cmdlen, int addr, int addrlen) {
+int32_t spi_setup_cmd_addr(SPI_TypeDef *spi, uint32_t cmd, uint32_t cmdlen, uint32_t addr, uint32_t addrlen) {
 
     if (__RARELY(spi == NULL)) {
         return -1;
@@ -18,7 +18,7 @@ int32_t spi_setup_cmd_addr(SPI_TypeDef *spi, int cmd, int cmdlen, int addr, int 
     return 0;
 }
 
-int32_t spi_setup_clk(SPI_TypeDef *spi, int clkdiv) {
+int32_t spi_setup_clk(SPI_TypeDef *spi, uint8_t clkdiv) {
 
     if (__RARELY(spi == NULL)) {
         return -1;
@@ -29,7 +29,7 @@ int32_t spi_setup_clk(SPI_TypeDef *spi, int clkdiv) {
     return 0;
 }
 
-int32_t spi_setup_dummy(SPI_TypeDef *spi, int dummy_rd, int dummy_wr) {
+int32_t spi_setup_dummy(SPI_TypeDef *spi, uint16_t dummy_rd, uint16_t dummy_wr) {
 
     if (__RARELY(spi == NULL)) {
         return -1;
@@ -40,7 +40,7 @@ int32_t spi_setup_dummy(SPI_TypeDef *spi, int dummy_rd, int dummy_wr) {
     return 0;
 }
 
-int32_t spi_set_datalen(SPI_TypeDef *spi, int datalen) {
+int32_t spi_set_datalen(SPI_TypeDef *spi, uint32_t datalen) {
 
     if (__RARELY(spi == NULL)) {
         return -1;
@@ -54,7 +54,7 @@ int32_t spi_set_datalen(SPI_TypeDef *spi, int datalen) {
     return 0;
 }
 
-int32_t spi_start_transaction(SPI_TypeDef *spi, int trans_type, int csnum) {
+int32_t spi_start_transaction(SPI_TypeDef *spi, uint32_t trans_type, uint32_t csnum) {
 
     if (__RARELY(spi == NULL)) {
         return -1;
@@ -74,7 +74,7 @@ int32_t spi_get_status(SPI_TypeDef *spi) {
     return spi->STATUS;
 }
 
-int32_t spi_write_fifo(SPI_TypeDef *spi, int *data, int datalen) {
+int32_t spi_write_fifo(SPI_TypeDef *spi, uint32_t *data, uint32_t datalen) {
 
     if (__RARELY(spi == NULL)) {
         return -1;
@@ -95,7 +95,7 @@ int32_t spi_write_fifo(SPI_TypeDef *spi, int *data, int datalen) {
     return 0;
 }
 
-int32_t spi_read_fifo(SPI_TypeDef *spi, int *data, int datalen) {
+int32_t spi_read_fifo(SPI_TypeDef *spi, uint32_t *data, uint32_t datalen) {
 
     if (__RARELY(spi == NULL)) {
         return -1;
